requirements.txt
fastapi[all]
psycopg2-binary
cryptography
pycryptodome
sqlalchemy
requests
jose
pyjwt
docker
numpy
pandas
python-dotenv
apache-airflow[postgres]
minio
pytorch-lightning
git+https://gitlab.com/PersonalHealthTrain/implementations/germanmii/difuture/train-container-library.git